% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_import}
\alias{data_import}
\title{asjdoiasdjaoisdjaij osaijdoias jdoiajsodijsaoi}
\usage{
data_import(file_name = "ebtrk_atlc_1988_2017.txt", path = "~")
}
\arguments{
\item{file_name}{The path to reach the file "ebtrk_atlc_1988_2015.txt".}
}
\value{
The outuput of this function is a data frame with storm_id, date, latitude,
        longitude, wind_speed, ne, nw, se, and sw.
}
\description{
asjdoiasdjaoisdjaij osaijdoias jdoiajsodijsaoi
}
