% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{The geom_hurricane function aims to plot a new form of data visualization, based on Wind Radius Speed
this geom will plot a graphich with "rose" format which will represent the area of the wind speed. All
plot must be created from a ggplot2 or ggmaps graphic.}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{Geom_harry_kane}{pirir}

\item{required_aes}{p irir}

\item{default_aes}{pirir}

\item{draw_key}{pirir}

\item{draw_group}{pirir}
}
\value{
The output of this geom will be a pie style of graphic varying the radius to show the intense
        of the wind speed.
}
\description{
The geom_hurricane function aims to plot a new form of data visualization, based on Wind Radius Speed
this geom will plot a graphich with "rose" format which will represent the area of the wind speed. All
plot must be created from a ggplot2 or ggmaps graphic.
}
\examples{

\dontrun{data_manipulation("ebtrk_atlc_1988_2015.txt")}

\dontrun{data_manipulation("ebtrk_atlc_1988_2017.txt")}

}
